@@ -14,14 +14,13 @@
   },
   "devDependencies": {
     "@fortawesome/fontawesome-free": "^6.0.0",
-    "@vue/test-utils": "^1.2.2",
+    "@vue/compat": "^3.1.0",
+    "@vue/compiler-sfc": "^3.1.0",
+    "@vue/test-utils": "^2.2.8",
+    "@vue/vue3-jest": "^28.1.0",
     "axios": "^0.21",
     "axios-mock-adapter": "^1.21.2",
-    "babel-core": "^7.0.0-0",
-    "babel-jest": "^27.2.1",
-    "bootstrap": "^4.6.0",
-    "bootstrap-vue": "^2.21.2",
-    "jest": "^27.2.1",
+    "babel-core": "^7.0.0-bridge.0",
     "jquery": "^3.6",
     "laravel-mix": "^6.0.6",
     "laravel-mix-valet": "^1.1.1",
@@ -31,20 +30,20 @@
     "resolve-url-loader": "^5.0.0",
     "sass": "^1.32.11",
     "sass-loader": "^11.0.1",
-    "vue": "^2.6.12",
-    "vue-jest": "^3.0.7",
-    "vue-loader": "^15.9.8",
-    "vue-template-compiler": "^2.6.12"
+    "vue": "^3.2.45",
+    "vue-loader": "^16.0.0"
   },
   "dependencies": {
+    "bootstrap": "^5.2.3",
+    "bootstrap-vue": "^2.23.1",
+    "jest-environment-jsdom": "^29.4.1",
     "moment": "^2.29.4",
     "numeral": "^2.0.6",
     "urijs": "^1.19.11",
-    "vue-clipboard2": "^0.3.3",
     "vue-multiselect": "^2.1.6",
-    "vue-router": "^3.5.2",
-    "vue-template-compiler": "^2.6.14",
-    "vuex": "^3.6.2"
+    "vue-router": "^4.0.0",
+    "vue3-clipboard": "^1.0.0",
+    "vuex": "^4.0.2"
   },
   "jest": {
     "moduleFileExtensions": [
@@ -53,7 +52,7 @@
       "vue"
     ],
     "transform": {
-      ".*\\.(vue)$": "vue-jest",
+      ".*\\.(vue)$": "@vue/vue-jest",
       "^.+\\.js$": "<rootDir>/node_modules/babel-jest"
     },
     "moduleNameMapper": {
@@ -62,6 +61,12 @@
     "setupFiles": [
       "<rootDir>/resources/js/tests/setup.js"
     ],
-    "testEnvironment": "jsdom"
+    "testEnvironment": "jsdom",
+    "testEnvironmentOptions": {
+      "customExportConditions": [
+        "node",
+        "node-addons"
+      ]
+    }
   }
 }
