@@ -1,18 +1,17 @@
-import Vue from 'vue';
-import Multiselect from 'vue-multiselect';
+// import Multiselect from 'vue-multiselect';
 
-Vue.component('clipboard', require('./components/shared/Clipboard').default);
+// app.component('clipboard', require('./components/shared/Clipboard').default);
 
-Vue.component('multiselect', Multiselect);
+// app.component('multiselect', Multiselect);
 
-Vue.component('nav-root', require('./components/nav/NavRoot').default);
+// app.component('nav-root', require('./components/nav/NavRoot').default);
 
-Vue.component('file-uploader', require('./components/shared/FileUploader').default);
+// app.component('file-uploader', require('./components/shared/FileUploader').default);
 
-Vue.component('field-error', require('./components/shared/FieldError').default);
+// app.component('field-error', require('./components/shared/FieldError').default);
 
-Vue.component('avatar', require('./components/home/ProfileImage').default);
+// app.component('avatar', require('./components/home/ProfileImage').default);
 
-Vue.component('page-footer', require('./components/shared/Footer').default);
+// app.component('page-footer', require('./components/shared/Footer').default);
 
-Vue.component('phone-number', require('./components/shared/PhoneNumber').default);
+// app.component('phone-number', require('./components/shared/PhoneNumber').default);
