@@ -20,7 +20,9 @@
 
     </head>
     <body>
-        <div id="app"><router-view style="min-height: 100vh;"></router-view></div>
+        <div id="app">
+            <router-view style="min-height: 100vh;"/>
+        </div>
         @include('modals.session-expired')
     <script>
         // Check that service workers are supported
