@@ -24,7 +24,7 @@ mix
   .js('resources/js/main.js', 'public/js')
   .vue()
   .sass('resources/sass/app.scss', 'public/css')
-  .sass('resources/sass/main.scss', 'public/css')
+  //.sass('resources/sass/main.scss', 'public/css')
   .copy('node_modules/@fortawesome/fontawesome-free/webfonts', 'public/webfonts');
 
 mix.version();
